Lots of UI, code, functionality changes
Adding a monopoly image to be used in the background of the game select and new game pages. Designing ui for game select and start menu. Starting very basics of the BoardController class, creating data classes cell, player and property. Changing the base information of the main application class. Making various methods in SQLUtils, Utils, and StartController to get the correct game information after submitting the game id and password. Adding a java method to change pages on the start page. Adding some basic styling and hover effects in a styles.css file.