Beginning functionality and organization, locking in
Continued adding ui information, drawCard method, and initialization to BoardController. Added previous position and goToJail in player. Removed the else statement in SQLUtils. Created classes for action cards and properties such as Action and Card. Changed privacy in Cell. Added an enum file to hold the state of every property. Various ui changes in both the board and start page.