My application is like Schrödinger's Cat, everything works and doesn't work simultaneously
Added button name parameters to confirmAlert in Utils. Converted cards to a local variable and made a single Action variable to hold the action in the card to make description and action name accurate and not random. Changed Go to an empty cell, added print statements to each action in SQLUtils. Added print statements and an inJail boolean in Player. Lots of changes in BoardController.