I am so done with monopoly
Basic board changes and added a back button to go from the game to the front page. Removed the owner of the property if setStage is called with a parameter of 0. Added spacing and sellProperty mechanics. The whole game should work except for game saving, it can load data properly but cannot save anything. Starting to design a ui for the cosmetology application with Mason.